Change recommended command to improve escaped character support (#63)
* Update README.md

Added step to install docker-compose

* Update README.md

* Update README.md

Changed command syntax for logging in to match syntax for registering because the syntax used for registering properly escapes bash special characters, but the syntax used for logging in does not. This allows you to create a user with a bash special character such as ! or | in the password, but the login command will fail with an error. The login works with special characters however if the same syntax is used as registration. I am updating this to hopefully help others avoid the same issue I discovered.